refactor(vod): split the 791-line manager.go by concern
Pure same-package file reorganization of the VOD Manager (no behavior change):
- manager.go (791 -> 160): Manager/types, lifecycle (New/Shutdown/CancelAll)
- metadata.go: the artifact metadata state machine (Mark*/Promote*/Demote*/prune)
- build.go:    build orchestration (StartBuild*/EnsureSpec*/Get)
- probe.go:    the probe path (Probe/enqueueProbe/TriggerProbe)

Build + go vet clean; vod tests green; golangci-lint --new-from-rev = 0; verify-config green.
